+++ src/contexts/ChainDataContext.tsx
@@ -29,6 +29,7 @@ interface ChainDataContextIF {
     setLastBlockNumber: Dispatch<SetStateAction<number>>;
     client: Client;
     connectedUserXp: UserXpDataIF;
+    isActiveNetworkBlast: boolean;
 }
 
 export const ChainDataContext = createContext<ChainDataContextIF>(
@@ -52,6 +53,13 @@ export const ChainDataContextProvider = (props: {
 
     const [lastBlockNumber, setLastBlockNumber] = useState<number>(0);
     const [gasPriceInGwei, setGasPriceinGwei] = useState<number | undefined>();
+
+    const isActiveNetworkBlast = [
+        '0x5d50',
+        '0xa0c71fd',
+        '0xee',
+        '0x13e31',
+    ].includes(chainData.chainId);
 
     async function pollBlockNum(): Promise<void> {
         // if default RPC is Infura, use key from env variable
@@ -214,6 +222,7 @@ export const ChainDataContextProvider = (props: {
         gasPriceInGwei,
         connectedUserXp,
         setGasPriceinGwei,
+        isActiveNetworkBlast,
         client,
     };
